added date/name/exercise comment and fixed hours and miles from int to double

+++ main.cs
@@ -1,5 +1,5 @@
 using System;
-
+//2023-09-15-AndrewHorton-Exercise2.2
 class Program {
   public static void Main (string[] args) {
     int baseRate = 200;
@@ -12,9 +12,9 @@ class Program {
      Console.WriteLine("Please enter the number of Miles ");
     var y = Console.ReadLine();
     
-    int totalHours;
-    int totalMiles;
-    if (int.TryParse(x, out totalHours) && int.TryParse(y, out totalMiles))
+    double totalHours;
+    double totalMiles;
+    if (double.TryParse(x, out totalHours) && double.TryParse(y, out totalMiles))
     {
       double totalCost = (baseRate) + ( perHourRate * totalHours ) + ( perMileRate * totalMiles);
       Console.WriteLine("${0}", totalCost );
